simply bootstrap script

+++ bootstrap.zsh
@@ -1,12 +1,12 @@
-#!/usr/bin/env zsh
+#!/usr/bin/env bash
 
-DIR="$( echo "$( dirname "$0" )" && pwd )"
 
 if [[ $SHELL != *"zsh"* ]];
 then
   echo "Installing oh-my-zsh"
   sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
 fi
+
 
 xcode-select -p > /dev/null
 if [ $? -ne 0 ];
@@ -18,13 +18,13 @@ fi
 if ! command -v brew > /dev/null;
 then
   echo "Installing Homebrew"
-  #/usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+  /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
 fi
+
 
 echo "Linking files"
 mkdir -p  ~/.dotfiles_backup
-for f in "aliases" "exports" "functions" "vimrc" "zshrc"
-do
+for f in "aliases" "exports" "functions" "vimrc" "zshrc"; do
   echo "Linking \"$f\""
   if [ -f ~/.$f ]; then
     echo "Original file exists, backing it up"
@@ -36,15 +36,22 @@ done
 
 
 # link Sublime Text
-if [ -f "/Applications/Sublime Text.app/Contents/SharedSupport/bin/subl" ]; then
+if [ -f "/Applications/Sublime Text.app/Contents/SharedSupport/bin/subl" ] \
+&& [ ! -e /usr/local/bin/subl ]; then
   ln -sv "/Applications/Sublime Text.app/Contents/SharedSupport/bin/subl" /usr/local/bin/subl
+else
+  echo "already exists, skipping"
 fi
 
-echo "Installing zsh-syntax-highlighting"
-brew install zsh-syntax-highlighting
-
-# Install vim + plugins
-brew install vim --with-lua
+software_list=( bash tig icdiff vim zsh-syntax-highlighting zsh-autosuggestions)
+for item in "${software_list[@]}"; do
+  echo "install or upgrading package: $item"
+  if [ -z $(brew list | grep "$item") ]; then
+    brew install "$item"
+  else
+    brew upgrade "$item" 2>/dev/null
+  fi
+done
 
 if [ ! -f ~/.vim/autoload/plug.vim ];
 then
@@ -55,7 +62,5 @@ fi
 
 echo "Configuring VIM"
 vim +PlugInstall +qall
-
-brew install tig icdiff
 
 echo "Done configuring the system, please reboot :D"
